+++ src/commandhandler/common.py
@@ -1,11 +1,34 @@
-#!/usr/bin/env python3
+##
+# Part of `SmartNodeMonitorBot`
+#
+# Copyright 2018 dustinface
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+##
 
 import logging
-from src import messages
-from src import util
 import requests
 import json
 import time
+
+from src import messages
+from src import util
 
 import telegram
 import discord
@@ -84,7 +107,7 @@ def info(bot, update):
                                               aberration)
 
         else:
-            response += "*Sorry, the bot is currently not synced with the network. Try it again in few minutes...*"
+            response += messages.notSynced(bot.messenger)
 
     return response
 
@@ -170,9 +193,6 @@ def payouts(bot, args):
     response += "E: {}%".format(round( (1 - (nList/total))*100, 1))
 
     return response
-
-
-
 
 ######
 # Telegram command handler for printing unknown command text