+++ src/util.py
@@ -1,4 +1,26 @@
-#!/usr/bin/env python3
+##
+# Part of `SmartNodeMonitorBot`
+#
+# Copyright 2018 dustinface
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+##
 
 import os, stat
 import threading
@@ -38,6 +60,11 @@ def secondsToText(secs):
     hours = (secs - days*86400)//3600
     minutes = (secs - days*86400 - hours*3600)//60
     seconds = secs - days*86400 - hours*3600 - minutes*60
+
+    if days:
+        seconds = -1
+        minutes = -1
+
     result =  ("{0} day{1}, ".format(days, "s" if days!=1 else "") if days else "")
     result += ("{0} hour{1}{2} ".format(hours, "s" if hours!=1 else "", "," if minutes and not days else "") if hours else "")
     result += ("{0} minute{1}{2} ".format(minutes, "s" if minutes!=1 else "", "," if seconds and not minutes else "") if minutes and not days else "")