+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - New node `history` command which shows past payouts for all nodes
 - Restrict the lookup command to DM
 - Show remaining wait time for nodes in the initial wait phase
+- New node `top` command to show only nodes in the top X percent